Make toImmutableEnum[Set|Map] to AndroidAccessToCollectors
I think that’s the last of them.

The implementations I moved verbatim, and there was no change of code there.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=342853260
